#43 Refactors code
