+++ Application/src/main/java/com/example/android/dancecar/Moves/Choreography.java
@@ -1,6 +1,4 @@
-package com.example.android.dancecar;
-
-import androidx.annotation.NonNull;
+package com.example.android.dancecar.Moves;
 
 import java.util.ArrayList;
 import java.util.Random;
@@ -17,29 +15,16 @@ public class Choreography {
         this.chorName = chorName;
     }
 
-
-
     public ArrayList<DanceMove> getSelectedDances() {
         return selectedDances;
-    }
-
-    public void setSelectedDances(ArrayList<DanceMove> selectedDances) {
-        this.selectedDances = selectedDances;
     }
 
     public int getChorMoveID() {
         return chorMoveID;
     }
 
-    public void setChorMoveID(int chorMoveID) {
-        this.chorMoveID = chorMoveID;
-    }
-
     public String getChorName() {
         return chorName;
     }
 
-    public void setChorName(String chorName) {
-        this.chorName = chorName;
-    }
 }
