+++ Application/src/main/java/com/example/android/dancecar/Moves/DanceMove.java
@@ -1,12 +1,8 @@
-package com.example.android.dancecar;
+package com.example.android.dancecar.Moves;
 
 import androidx.annotation.NonNull;
 
 import java.util.Random;
-import java.util.UUID;
-import java.util.Random.*;
-
-import java.util.ArrayList;
 
 public class DanceMove {
     private String danceName;
@@ -26,13 +22,8 @@ public class DanceMove {
         return danceName;
     }
 
-
     public String getDanceName() {
         return danceName;
-    }
-
-    public void setDanceName(String danceName) {
-        this.danceName = danceName;
     }
 
     public int getId() {
@@ -41,14 +32,6 @@ public class DanceMove {
 
     public void setId(int id) {
         this.id = id;
-    }
-
-    public Random getRandom() {
-        return random;
-    }
-
-    public void setRandom(Random random) {
-        this.random = random;
     }
 
     public boolean isCreated() { return isCreated; }
