+++ Application/src/main/res/layout/activity_dancing.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fadeScrollbars="false"
-    tools:context=".DancingActivity">
+    tools:context=".Activities.DancingActivity">
 
 
     <ImageView
